+++ src/torchmetrics/__init__.py
@@ -43,7 +43,7 @@ from torchmetrics.classification import (  # noqa: E402
     StatScores,
 )
 from torchmetrics.collections import MetricCollection  # noqa: E402
-from torchmetrics.detection import PanopticQuality  # noqa: E402
+from torchmetrics.detection import ModifiedPanopticQuality, PanopticQuality  # noqa: E402
 from torchmetrics.image import (  # noqa: E402
     ErrorRelativeGlobalDimensionlessSynthesis,
     MultiScaleStructuralSimilarityIndexMeasure,
@@ -153,6 +153,7 @@ __all__ = [
     "MetricTracker",
     "MinMaxMetric",
     "MinMetric",
+    "ModifiedPanopticQuality",
     "MultioutputWrapper",
     "MultiScaleStructuralSimilarityIndexMeasure",
     "PanopticQuality",