+++ src/torchmetrics/regression/mae.py
@@ -55,6 +55,8 @@ class MeanAbsoluteError(Metric):
     is_differentiable: bool = True
     higher_is_better: bool = False
     full_state_update: bool = False
+    plot_lower_bound: float = 0.0
+
     sum_abs_error: Tensor
     total: Tensor