+++ src/torchmetrics/retrieval/r_precision.py
@@ -78,6 +78,8 @@ class RetrievalRPrecision(RetrievalMetric):
     is_differentiable: bool = False
     higher_is_better: bool = True
     full_state_update: bool = False
+    plot_lower_bound: float = 0.0
+    plot_upper_bound: float = 1.0
 
     def _metric(self, preds: Tensor, target: Tensor) -> Tensor:
         return retrieval_r_precision(preds, target)