+++ src/torchmetrics/wrappers/classwise.py
@@ -11,11 +11,16 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, Dict, List, Optional, Sequence, Union
 
 from torch import Tensor
 
 from torchmetrics import Metric
+from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE
+from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE
+
+if not _MATPLOTLIB_AVAILABLE:
+    __doctest_skip__ = ["ClasswiseWrapper.plot"]
 
 
 class ClasswiseWrapper(Metric):
@@ -114,3 +119,46 @@ class ClasswiseWrapper(Metric):
     def _wrap_compute(self, compute: Callable) -> Callable:
         """Overwrite to do nothing."""
         return compute
+
+    def plot(
+        self, val: Optional[Union[Tensor, Sequence[Tensor]]] = None, ax: Optional[_AX_TYPE] = None
+    ) -> _PLOT_OUT_TYPE:
+        """Plot a single or multiple values from the metric.
+
+        Args:
+            val: Either a single result from calling `metric.forward` or `metric.compute` or a list of these results.
+                If no value is provided, will automatically call `metric.compute` and plot that result.
+            ax: An matplotlib axis object. If provided will add plot to that axis
+
+        Returns:
+            Figure and Axes object
+
+        Raises:
+            ModuleNotFoundError:
+                If `matplotlib` is not installed
+
+        .. plot::
+            :scale: 75
+
+            >>> # Example plotting a single value
+            >>> import torch
+            >>> from torchmetrics import ClasswiseWrapper
+            >>> from torchmetrics.classification import MulticlassAccuracy
+            >>> metric = ClasswiseWrapper(MulticlassAccuracy(num_classes=3, average=None))
+            >>> metric.update(torch.randint(3, (20,)), torch.randint(3, (20,)))
+            >>> fig_, ax_ = metric.plot()
+
+        .. plot::
+            :scale: 75
+
+            >>> # Example plotting multiple values
+            >>> import torch
+            >>> from torchmetrics import ClasswiseWrapper
+            >>> from torchmetrics.classification import MulticlassAccuracy
+            >>> metric = ClasswiseWrapper(MulticlassAccuracy(num_classes=3, average=None))
+            >>> values = [ ]
+            >>> for _ in range(3):
+            ...     values.append(metric(torch.randint(3, (20,)), torch.randint(3, (20,))))
+            >>> fig_, ax_ = metric.plot(values)
+        """
+        return self._plot(val, ax)